read pod env variables
